pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vacancies-parser"
version = "0.1.0"
description = "Hourly ML vacancy scraper with Telegram digests"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "httpx>=0.27.0",
  "lxml>=5.2.0",
  "playwright>=1.51.0",
  "psycopg[binary]>=3.2.0",
  "PyYAML>=6.0.0",
  "SQLAlchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
packages = ["app", "app.parsers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
